requirements.txt
schedule
requests 
python-telegram-bot